admin can edit only

+++ app/src/main/java/com/example/tianhao/seg2105project/Model/Application.java
@@ -38,10 +38,7 @@ public class Application {
                 serviceArrayList = new ArrayList<>();
                 for(DataSnapshot data : dataSnapshot.getChildren()){
                     serviceArrayList.add(data.getValue(Service.class));
-
-//                    Toast.makeText(mContext, "lululu", Toast.LENGTH_SHORT).show();
                 }
-//                adapter = new ServiceViewAdapter(mContext,serviceArrayList);
             }
 
             @Override
@@ -59,14 +56,14 @@ public class Application {
     }
 
     public void addService(String name, double hourlyRate){
-//        if(user.getUserType()!="Administrator") {return;}
+        if(user.getUserType()!="Administrator") {return;}
         String id = services.push().getKey();
         services.child(id).setValue(new Service(id,name,hourlyRate));
         Toast.makeText(mContext, "New Service Added", Toast.LENGTH_SHORT).show();
     }
 
     public void editService(String id, String name, double hourlyRate){
-//        if(user.getUserType()!="Administrator") {return;}
+        if(user.getUserType()!="Administrator") {return;}
         DatabaseReference originalService = services.child(id);
         originalService.setValue(new Service(id, name, hourlyRate));
         Toast.makeText(mContext, "Service Updated", Toast.LENGTH_SHORT).show();
@@ -74,7 +71,7 @@ public class Application {
 
 
     public void removeService(String id){
-//        if(user.getUserType()!="Administrator") {return;}
+        if(user.getUserType()!="Administrator") {return;}
         if(id.equals(""))return;
         DatabaseReference originalService = services.child(id);
         originalService.removeValue();
